+++ example/lib/features/image_format_convert_example.dart
@@ -73,7 +73,10 @@ class _ImageFormatConvertExampleState extends State<ImageFormatConvertExample>
 
           setGenerationTime();
         },
-        onCloseEditor: () => onCloseEditor(enablePop: !isDesktopMode(context)),
+        onCloseEditor: (editorMode) => onCloseEditor(
+          editorMode: editorMode,
+          enablePop: !isDesktopMode(context),
+        ),
         mainEditorCallbacks: MainEditorCallbacks(
           helperLines: HelperLinesCallbacks(onLineHit: vibrateLineHit),
         ),