+++ example/lib/features/video_examples/mixins/video_editor_mixin.dart
@@ -155,7 +155,8 @@ mixin VideoEditorMixin<T extends StatefulWidget> on State<T> {
   ///
   /// If [exportedVideo] is available, it navigates to [PreviewVideo].
   /// Afterwards, it pops the current editor page.
-  void onCloseEditor() async {
+  void onCloseEditor(EditorMode editorMode) async {
+    if (editorMode != EditorMode.main) return Navigator.pop(context);
     if (exportedVideo != null) {
       await Navigator.push(
         context,